pyproject.toml
[project]
name = "testing-goat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ansible>=12.2.0",
    "django>=5.2.8",
    "django-stubs[compatible-mypy]>=5.2.7",
    "docker>=7.1.0",
    "gunicorn>=23.0.0",
    "lxml[cssselect]>=6.0.2",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "pytest-playwright>=0.7.1",
    "ruff>=0.14.3",
    "whitenoise>=6.11.0",
]


[tool.django-stubs]
django_settings_module = "superlists.settings"


[tool.mypy]
plugins = ["mypy_django_plugin.main"]


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "superlists.settings"
python_files = ["tests.py", "tests/*.py", "test_*.py"]

